Add return structure for SDK callback
To avoid sending several params I made a structure to hold
all possible information.
All that is required is a result code describing the current state.
It contains one of the following states:
    UPDATE_REQUIRED,
    UPDATE_AVAILABLE,
    UP_TO_DATE,
    ERROR

It is up to the developer to decide what is the appropriate action
to take in each case.