Add alert dialog to notify user of new versions
If the result returns UPDATE_AVAILABLE a dialog with
Download or Dismiss will show.
If the result is UPDATE_REQUIRED the Dismiss button is
removed and the dialog activity locks down the app until the
required version has been installed.
Clicking Download will route you to the url received from the server.

This class if optional to implement and we encourage the developer
to implement his or her own way of handling version conflicts.
This is just a tool to get you started.