Change comparison from version name to version code
Restructured a large amount of the code base to change the comparison
from the version name string value to the version code Int value.
As a result i had to change the api models and subsequently all code based on
it.

A lot of code could be removed since code comparison is less challenging then
first convert a string to an int and then compare it to another string.
This also means a lot of the test became obsolete and was cut.